Camera FIFO driver model

+++ sw/skynet/src/motor_driver.c
@@ -7,9 +7,10 @@
 
 #include "motor_driver.h"
 
-void SetDutyCycle(void* base, uint32_t dutyCycle, uint32_t dir)
+void SetDutyCycle(void* base, float dutyCycle, uint32_t dir)
 {
-	uint32_t mod = (dutyCycle / 100) * 255;
+	// This will truncate the float, this is expected
+	uint32_t mod = (dutyCycle / 100.0) * 255.0; 
 	MOTOR(base)->DIR = dir;
 	MOTOR(base)->DUTY_CYCLE = mod;
 }
